+++ mcc/lexer.py
@@ -10,8 +10,8 @@ from mcc.errors import LangError
 KEYWORDS = {"fn", "return", "let", "const", "if", "else", "while", "until",
             "break", "continue", "defer", "for", "in", "case", "when", "and",
             "or", "true", "false", "import", "as", "sizeof", "alignof",
-            "offsetof", "len", "struct", "extends", "null", "emit", "enum",
-            "union", "mut", "unreachable"}
+            "offsetof", "typename", "len", "struct", "extends", "null", "emit",
+            "enum", "union", "mut", "unreachable"}
 
 TOKEN_SPEC = [
     ("COMMENT", r"//[^\n]*|/\*(?s:.*?)\*/"),